refactored redux folder
