+++ src/pages/cart/index.js
@@ -1,7 +1,7 @@
 import { useCart } from "../context";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
-import { selectCurrentUser } from "../../auth/authSelector";
+import { selectCurrentUser } from "../../redux/authSelector";
 import CustomButton from "../../components/button/index";
 import NavScrollExample from "../../layout/navbar/index";
 import { FaTrash } from "react-icons/fa6";
