+++ src/pages/login/style.scss
@@ -2,20 +2,20 @@
 
 .login-form-container {
   background-image: var(--bg-url);
-  background-size: cover;
+  background-size: 85%;
   background-position: center;
   min-height: 100vh;
   display: flex;
   justify-content: flex-start;
   align-items: center;
-  padding-left: 5rem;
-  padding-right: 2rem;
+  padding: 2rem 5rem;
 }
 
 .login-content {
+  background-color: $text-light;
   padding: 2rem;
   border-radius: 10px;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.203);
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.47);
   width: 100%;
   max-width: 400px;
   display: flex;
@@ -26,8 +26,9 @@
 .section-heading {
   font-size: 2.5rem;
   font-weight: 700;
-  color: #333333;
+  color: $primary-color;
   margin-bottom: 1rem;
+  text-align: center;
 }
 
 .input-section {
