+++ src/redux/reducers.js
@@ -1,0 +1,1 @@
+export { default as authReducer, logout } from "./user/reducers";
