+++ src/redux/store/index.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import authReducer from "../../auth/authSlice";
+import { authReducer } from "../reducers";
 
 export const store = configureStore({
   reducer: {
